ui/requirements.txt
flask==2.3.3
pywebview==4.0.1
Pillow==10.0.0
requests==2.32.4
urllib3==2.5.0
certifi==2025.8.3
